feat: 完善上传文件

+++ shared/src/commonMain/kotlin/com/lemon/mcdevmanagermp/data/vo/netease/upload/UploadFileVO.kt
@@ -8,7 +8,7 @@ data class UploadFileVO(
     val url: String,
     val mime: String,
     @SerialName("fsize")
-    val fileSize: Int,
+    val fileSize: Long,
     val md5: String,
     val picSize: List<Int>? = null
 )
